+++ src/Components/Checkbox/index.ts
@@ -1,0 +1,2 @@
+export { default as Checkbox } from "./Checkbox"
+export type { CheckBoxProps } from "./Checkbox"
