+++ src/Components/DateRange/index.ts
@@ -1,0 +1,2 @@
+export { default as DateRange } from "./DateRange"
+export type { DateRangeProps } from "./DateRange"
