+++ src/Components/HeadCell/index.ts
@@ -1,0 +1,2 @@
+export { default as HeadCell } from "./HeadCell"
+export type { HeadCellProps } from "./HeadCell"
