+++ src/Components/NumberField/index.ts
@@ -1,0 +1,2 @@
+export { default as NumberField } from "./NumberField"
+export type { NumberFieldProps, NumberType } from "./NumberField"
