+++ src/Components/NumberRange/index.ts
@@ -1,0 +1,2 @@
+export { default as NumberRange } from "./NumberRange"
+export type { NumberRangeProps } from "./NumberRange"
