+++ src/Components/SearchField/index.ts
@@ -1,0 +1,2 @@
+export { default as SearchField } from "./SearchField"
+export type { SearchFieldProps } from "./SearchField"
