+++ src/Components/SelectList/index.ts
@@ -1,0 +1,2 @@
+export { default as SelectList } from "./SelectList"
+export type { SelectListItem, SelectListProps } from "./SelectList"
