+++ src/Components/TextField/index.ts
@@ -1,0 +1,2 @@
+export { default as TextField } from "./TextField"
+export type { TextFieldProps } from "./TextField"
