+++ src/Components/Toolbar/index.tsx
@@ -1,4 +1,5 @@
 export { default as ToolbarButton } from "./ToolbarButton"
-export { default as CrudToolbar } from "./CrudToolbar"
-
 export type { ToolbarButtonProps } from "./ToolbarButton"
+
+export { default as CrudToolbar } from "./CrudToolbar"
+export type { CrudToolbarButtonProps, CrudToolbarProps, AdditionalToolbarProps, ToolbarProps } from "./CrudToolbar"
