+++ src/DataTable/hooks/index.ts
@@ -1,3 +1,5 @@
+import { FourMpRounded } from "@mui/icons-material"
+
 export { default as useSelectRows } from "./useSelectRows"
 export { default as useSelectAllStatus } from "./useSelectAllStatus"
 export { default as usePrevious } from "./usePrevious"
@@ -5,4 +7,6 @@ export { default as useSorting } from "./useSorting"
 export { default as useSearching } from "./useSearching"
 export { default as useFilter } from "./useFilter"
 
-export { SELECT_ALL_STATUSES } from "./useSelectAllStatus"
+export type {
+  FilterChecker, FilterCheckers, FilterComparer, FilterComparers, FilterProps
+} from "./types"
