+++ src/DataTable/index.ts
@@ -1,3 +1,9 @@
 export * from "./hooks"
-export * from "./DataTable"
+export { default as DataTable } from "./DataTable"
 export * from "./defaultProps"
+
+export * from "./hooks"
+
+export type {
+  BodyLineCell, DataTableProps, HeadLineCell, TableRowProps, LineCell
+} from "./types"
