+++ src/Filter/index.ts
@@ -1,4 +1,7 @@
 export * from "./hooks"
 export * from "./BooleanFilter"
-export * from "./Filter"
+export { default as Filter } from "./Filter"
 export * from "./FilterContainer"
+export * from "./hooks"
+
+export * from "./types"
