+++ src/index.ts
@@ -1,5 +1,5 @@
-export { default as DataTable } from "./DataTable/DataTable";
-
-export type {
-  BodyLineCell, DataTableProps, LineCell, HeadLineCell, TableRowProps
-} from "./DataTable/DataTable";
+export * from "./DataTable";
+// export * from "./Filter";
+export * from "./Components";
+export * from "./const";
+export * from "./utils";
